Relocate structs to another file
