+++ pkg/cmd/webhooks/structs.go
@@ -1,6 +1,11 @@
 package webhooks
 
-import "net/http"
+import (
+	"net/http"
+	"sync"
+
+	"github.com/gorilla/websocket"
+)
 
 type createHookRequest struct {
 	Name   string     `json:"name"`
@@ -34,4 +39,23 @@ type httpEventForward struct {
 	Status int
 	Header http.Header
 	Body   []byte
+}
+
+type ConnCloser struct {
+	*websocket.Conn
+	mu     sync.Mutex
+	closed bool
+}
+
+func (c *ConnCloser) Close() error {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	c.closed = true
+	return c.Conn.Close()
+}
+
+func (c *ConnCloser) IsClosed() bool {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	return c.closed
 }
